chore: relocate mobile assets
