+++ apps/mobile/src/assets/images/mocks/ape.png
