+++ apps/mobile/src/assets/images/mocks/hidetaka.png
