+++ apps/mobile/src/assets/images/platform-icons/adaptive-icon.png
