+++ apps/mobile/src/assets/images/platform-icons/favicon.png
