+++ apps/mobile/src/assets/images/platform-icons/icon.png
