+++ apps/mobile/src/assets/images/socials/email.png
