+++ apps/mobile/src/assets/images/socials/sns-discord.png
