+++ apps/mobile/src/assets/images/socials/sns-twitter.png
