+++ apps/mobile/src/assets/images/socials/telegram.png
